Update decred-101.vue
Overall great start. I'd like to see some updates to the intro but tried to keep it to just grammar/spelling/sentence structure. There are some inaccuracies that need to be discussed/I edited a few. This is the most important page and needs to be more of a summary than current. Examples/additional thoughts:

1. Banks printing money - we may want to distinguish commercial banks and the Fed
2. Pseudorandom time lock up for staking (I edited) - it is not "unknown"
3. I think we present the cost to attack calculator later in the articles - too technical for the 101 page
4. There is sporadic mention of the reward system with no mention that stakeholders get rewarded through the tickets/PoS system - maybe we make more sense of what the block reward is as well. 
5. Maybe we put a link around "blockchain" to a blockchain 101 for those who really are new to the space and the system. Permabull has a lot of good blockchain 101 language we could leverage if we wanted to create our own. 

I understand this is meant to be a rough first draft but am happy to take a stab at editing actual content as well as I think there is still some content and structure work to be done. Let me know your thoughts - not trying to add work - just wanting to give as honest of a review as possible as I believe this has the potential to add tremendous value to the project!